refactor(gui): move filter components to new feature
